+++ src/widgets/programs/program-record.css
@@ -1,0 +1,34 @@
+:global(.omrs-breakpoint-gt-phone) .programSummary {
+  display: flex;
+}
+
+:global(.omrs-breakpoint-lt-tablet) .programSummary {
+  padding: 0.25rem 0.5rem 0 0.5rem;
+}
+
+.programCard {
+  margin: 0rem 1rem;
+}
+
+.programTable {
+  width: 100%;
+  margin-top: 1rem;
+  margin-bottom: 1rem;
+}
+
+.programTable thead th {
+  color: var(--omrs-color-ink-medium-contrast);
+  text-align: left;
+}
+
+.programTable thead {
+  color: var(--omrs-color-ink-medium-contrast);
+}
+
+.programTable tbody td:nth-child(1) {
+  font-family: "Work Sans";
+}
+
+.programTable thead td {
+  margin-top: 0.625rem;
+}
