+++ src/widgets/programs/programs-detailed-summary.css
@@ -1,34 +1,56 @@
-:global(.omrs-breakpoint-gt-phone) .programSummary {
+:global(.omrs-breakpoint-gt-phone) .programsSummary {
   display: flex;
+  justify-content: center;
 }
 
-:global(.omrs-breakpoint-lt-tablet) .programSummary {
+:global(.omrs-breakpoint-lt-tablet) .programsSummary {
   padding: 0.25rem 0.5rem 0 0.5rem;
-}
-
-.programCard {
-  margin: 0rem 1rem;
 }
 
 .programTable {
   width: 100%;
-  margin-top: 1rem;
-  margin-bottom: 1rem;
-}
-
-.programTable thead th {
-  color: var(--omrs-color-ink-medium-contrast);
-  text-align: left;
 }
 
 .programTable thead {
   color: var(--omrs-color-ink-medium-contrast);
-}
-
-.programTable tbody td:nth-child(1) {
-  font-family: "Work Sans";
+  border-bottom: 0.0625rem solid var(--omrs-color-bg-low-contrast);
 }
 
 .programTable thead td {
-  margin-top: 0.625rem;
+  padding: 1rem 0.25rem 1rem 1rem;
+}
+
+.programTable tbody {
+  padding: 2rem;
+  border-bottom: 0.0625rem solid var(--omrs-color-bg-low-contrast);
+}
+
+:global(.omrs-breakpoint-lt-tablet) .programTable thead td {
+  padding: 0.25rem 0.125rem 0.5rem 0.5rem;
+}
+
+.programsSummary tr {
+  padding: 0.813rem 0 1rem 0.813rem;
+  border-bottom: 0.0625rem solid var(--omrs-color-bg-low-contrast);
+}
+
+.programTable td {
+  padding: 0.75rem 0 0.75rem 1rem;
+}
+
+.programTable tbody td:nth-child(2) {
+  font-family: "Work Sans";
+}
+
+.programTable tbody td:nth-child(3) {
+  font-family: "Work Sans";
+}
+
+tr.inactive {
+  background: var(--omrs-color-bg-medium-contrast);
+  color: var(--omrs-color-ink-medium-contrast);
+}
+
+tr.active {
+  background: var(--omrs-color-bg-high-contrast);
 }
