--- src/widgets/programs/programs-summary.css
@@ -1,56 +1,0 @@
-:global(.omrs-breakpoint-gt-phone) .programsSummary {
-  display: flex;
-  justify-content: center;
-}
-
-:global(.omrs-breakpoint-lt-tablet) .programsSummary {
-  padding: 0.25rem 0.5rem 0 0.5rem;
-}
-
-.programTable {
-  width: 100%;
-}
-
-.programTable thead {
-  color: var(--omrs-color-ink-medium-contrast);
-  border-bottom: 0.0625rem solid var(--omrs-color-bg-low-contrast);
-}
-
-.programTable thead td {
-  padding: 1rem 0.25rem 1rem 1rem;
-}
-
-.programTable tbody {
-  padding: 2rem;
-  border-bottom: 0.0625rem solid var(--omrs-color-bg-low-contrast);
-}
-
-:global(.omrs-breakpoint-lt-tablet) .programTable thead td {
-  padding: 0.25rem 0.125rem 0.5rem 0.5rem;
-}
-
-.programsSummary tr {
-  padding: 0.813rem 0 1rem 0.813rem;
-  border-bottom: 0.0625rem solid var(--omrs-color-bg-low-contrast);
-}
-
-.programTable td {
-  padding: 0.75rem 0 0.75rem 1rem;
-}
-
-.programTable tbody td:nth-child(2) {
-  font-family: "Work Sans";
-}
-
-.programTable tbody td:nth-child(3) {
-  font-family: "Work Sans";
-}
-
-tr.inactive {
-  background: var(--omrs-color-bg-medium-contrast);
-  color: var(--omrs-color-ink-medium-contrast);
-}
-
-tr.active {
-  background: var(--omrs-color-bg-high-contrast);
-}
